chore: refactored src/utils/fetchAttestations.ts
separated concerns in different files by functions and types.
NOTE: this change does not affect the codebase as everything that was exported by fetchAttestations.ts now is exported by src/utils/fetchAttestations/index.ts
